moved view classes to main package to resolve preview issue
