+++ enhanced-edittext/src/main/java/fr/dvilleneuve/android/widget/EnhancedButton.java
@@ -16,12 +16,14 @@
  * It uses Iconify by Joan Zapata, licensed under Apache License version 2, which is compatible
  * with this library's license.
  */
-package android.widget;
+package fr.dvilleneuve.android.widget;
 
 import android.content.Context;
 import android.content.res.ColorStateList;
 import android.graphics.drawable.Drawable;
 import android.util.AttributeSet;
+import android.widget.Button;
+
 import com.joanzapata.iconify.Icon;
 import fr.dvilleneuve.android.EnhancedText;
 import fr.dvilleneuve.android.OnClickDrawableListener;
